Developped a working schema, and working insertion query. Currently testing 10K insertion by 4 workers => 40K insertion.
