+++ package.json
@@ -15,14 +15,16 @@
   "homepage": "https://github.com/bamboo-connection/map-side-bar#readme",
   "scripts": {
     "nosql-db": "mongod",
-    "sql-db": "psql -U ellisona -a -f db-sql/controllers/schema.sql",
+    "sql-db": "psql -U ellisona -a -f db-sql/models/schema.sql",
     "OG-seed": "node db/controllers/seed_database.js",
-    "seed": "node --max-old-space-size=3072 db/controllers/seed-tenM-NoSQL.js",
+    "nosql-seed": "node --max-old-space-size=3072 db-nosql/controllers/seed-tenM-NoSQL.js",
+    "sql-seed": "node --max-old-space-size=3072 db-sql/controllers/seed-tenM-SQL.js",
     "test": "jest",
     "react-dev": "webpack --watch",
     "server:watch": "nodemon server/index.js",
     "start": "node server/index.js",
-    "docker-start": "DATABASE_HOST=database node db/controllers/seed_database.js && DATABASE_HOST=database node server/index.js"
+    "sql-docker-start": "DATABASE_HOST=database node db-sql/controllers/seed_database.js && DATABASE_HOST=database node server/index.js",
+    "nosql-docker-start": "DATABASE_HOST=database node db-nosql/controllers/seed_database.js && DATABASE_HOST=database node server/index.js"
   },
   "devDependencies": {
     "babel-core": "^6.26.0",
